add delete product, discounts, and launcher icon

+++ feature/sales/src/main/java/com/sonder/simpleerp/sales/productsList/bottomsheet/AddProductBottomSheet.kt
@@ -70,7 +70,7 @@ class AddProductBottomSheet : BottomSheetDialogFragment() {
 
     private fun getProductName() = binding.addProductName.text.toString()
 
-    private fun getValidQuantity() = binding.addProductQuantity.text?.toString()?.toInt() ?: 0
+    private fun getValidQuantity() = binding.addProductQuantity.text?.toString()?.toFloat() ?: 0.0f
 
     private fun setupObservers() {
         viewLifecycleOwner.lifecycleScope.launch {
